Absences avec validation et choix des destinataires des notifications (Tous, responsable direct, agents ayant le droit de gérer les absences)

+++ absences/modif.php
@@ -7,7 +7,7 @@ Copyright (C) 2011-2013 - Jérôme Combes
 
 Fichier : absences/modif.php
 Création : mai 2011
-Dernière modification : 31 octobre 2013
+Dernière modification : 25 novembre 2013
 Auteur : Jérôme Combes, jerome@planningbilbio.fr
 
 Description :
@@ -21,6 +21,7 @@ require_once "class.absences.php";
 //	Initialisation des variables
 $display=null;
 $checked=null;
+$admin=in_array(1,$droits)?true:false;
 
 $id=$_GET['id'];
 $db=new db();
@@ -28,6 +29,7 @@ $req="SELECT `{$dbprefix}personnel`.`id` AS `perso_id`, `{$dbprefix}personnel`.`
   `{$dbprefix}personnel`.`prenom` AS `prenom`, `{$dbprefix}absences`.`id` AS `id`, `{$dbprefix}personnel`.`site` AS `site`,  
   `{$dbprefix}absences`.`debut` AS `debut`, `{$dbprefix}absences`.`fin` AS `fin`, 
   `{$dbprefix}absences`.`nbjours` AS `nbjours`, `{$dbprefix}absences`.`motif` AS `motif`, 
+  `{$dbprefix}absences`.`valide` AS `valide`, `{$dbprefix}absences`.`validation` AS `validation`, 
   `{$dbprefix}absences`.`commentaires` AS `commentaires` 
   FROM `{$dbprefix}absences` INNER JOIN `{$dbprefix}personnel` 
   ON `{$dbprefix}absences`.`perso_id`=`{$dbprefix}personnel`.`id` WHERE `{$dbprefix}absences`.`id`='$id';";
@@ -38,6 +40,8 @@ $commentaires=$db->result[0]['commentaires'];
 $debut=$db->result[0]['debut'];
 $fin=$db->result[0]['fin'];
 $site=$db->result[0]['site'];
+$valide=$db->result[0]['valide'];
+$validation=$db->result[0]['validation'];
 $hre_debut=substr($debut,-8);
 $hre_fin=substr($fin,-8);
 $debut=substr($debut,0,10);
@@ -46,6 +50,11 @@ if($hre_debut=="00:00:00" and $hre_fin=="23:59:59"){
   $checked="checked='checked'";
   $display="style='display:none;'";
 }
+$select1=$valide>0?"selected='selected'":null;
+$select2=$valide<0?"selected='selected'":null;
+$validation_texte=$valide>0?"Valid&eacute;e":"&nbsp;";
+$validation_texte=$valide<0?"Refus&eacute;e":$validation_texte;
+$validation_texte=$valide==0?"Demand&eacute;e":$validation_texte;
 
 // Sécurité
 // Droit 1 = modification de toutes les absences
@@ -129,28 +138,50 @@ echo "<option value=''>-----------------------</option>\n";
 $db_select=new db();
 $db_select->query("select valeur from {$dbprefix}select_abs order by rang;");
 foreach($db_select->result as $elem){
-  echo "<option value='".$elem['valeur']."'>".$elem['valeur']."</option>\n";
+  $selected=$elem['valeur']==$motif?"selected='selected'":null;
+  echo "<option value='".$elem['valeur']."' $selected>".$elem['valeur']."</option>\n";
 }
 echo "</select>\n";
-echo "<a href='javascript:popup(\"include/ajoutSelect.php&amp;table=select_abs&amp;terme=motif\",400,400);'>\n";
-echo "<img src='img/add.gif' alt='*' style=width:15px;'/></a>\n";
+if($admin){
+  echo "<a href='javascript:popup(\"include/ajoutSelect.php&amp;table=select_abs&amp;terme=motif\",400,400);'>\n";
+  echo "<img src='img/add.gif' alt='*' style=width:15px;'/></a>\n";
+}
 
-
-echo "</td></tr><tr><td>";
+echo "</td></tr><tr style='vertical-align:top;'><td>";
 echo "Commentaires : </td><td>";
 echo "<textarea name='commentaires' cols='25' rows='5'>$commentaires</textarea>";
 echo "</td></tr>";
+
+if($config['Absences-validation']){
+  echo "<tr><td>Validation : </td><td>\n";
+  if($admin){
+    echo "<select name='valide'>\n";
+    echo "<option value=''>&nbsp;</option>\n";
+    echo "<option value='1' $select1>Accept&eacute;e</option>\n";
+    echo "<option value='-1' $select2>Refus&eacute;e</option>\n";
+    echo "</select>\n";
+  }
+  else{
+    echo $validation_texte;
+  }
+  echo "</td></tr>\n";
+}
+
 echo "<tr><td colspan='2'><br/>\n";
-echo "<input type='button' value='Supprimer' onclick='document.location.href=\"index.php?page=absences/delete.php&amp;id=$id\";'/>";
-echo "&nbsp;&nbsp;\n";
-echo "<input type='button' value='Annuler' onclick='annuler(1);'/>\n";
-echo "&nbsp;&nbsp;\n";
-echo "<input type='submit' value='Valider'/>\n";
+if($admin or $valide==0 or $config['Absences-validation']==0){
+  echo "<input type='button' value='Supprimer' onclick='document.location.href=\"index.php?page=absences/delete.php&amp;id=$id\";'/>";
+  echo "&nbsp;&nbsp;\n";
+  echo "<input type='button' value='Annuler' onclick='annuler(1);'/>\n";
+  echo "&nbsp;&nbsp;\n";
+  echo "<input type='submit' value='Valider'/>\n";
+}
+else{
+  echo "<input type='button' value='Retour' onclick='annuler(1);'/>\n";
+}
 echo "</td></tr>\n";
 echo "</table>\n";
 echo "<input type='hidden' name='id' value='$id'/>";
 echo "</form>\n";
-echo "<script type='text/JavaScript'>initform(\"motif=$motif\");</script>";
 echo "<script type='text/JavaScript'>initform(\"hre_debut=$hre_debut\");</script>";
 echo "<script type='text/JavaScript'>initform(\"hre_fin=$hre_fin\");</script>";
 ?>
